Mahalanobis Distance class between two representations is added
until here, only the calculation for choosing the point of the biggest distance is added, i.e. the loss for fitting cannot be calculated yet

right after the seed point was added, the data pool is loaded to get all representations to compute the covariance

if the target set is used as input, still only the target set, will be considered for the covariance matrix, not the full data set.