Create new observables in jax_recording.py
- covariance entropy
- magnitude entropy

The covariance entropy is the von Neumann entropy of the real covariance NTK which is the re-scaled NTK such that it only carries cosine similarities.

The magnitude entropy is the shannon entropy of the density vector of gradient magnitudes.
The magnitude density is calculated via a the compute_magnitude density function in matrix_utils.py.

The covariance NTK is not available in the recording process, as it can be derived from the ntk by applying one function.