Fixing the slowdown in partitioned_training.py when using RND.
Identifying the issue as the following:
The RND build_data_set method hands the batch size to the train_model methdo fo the training strategy.
In the case of partitioned_training.py this is a list of batch_sizes.
When calling the training for the first time, the target set consists of a single point. Due to the training decorator, the batch_sizes get adapted to a list of ones. This however, overwrites the list in the RND agent, which will remain to be a list of ones, as it does not get updated afterwards. The training decorator only updates the batch size under certain conditions and none of them is fulfilled for batch sizes of ones.