Konsti debug partitioned training (#85)
* Fixing the slowdown in partitioned_training.py when using RND.

Identifying the issue as the following:
The RND build_data_set method hands the batch size to the train_model methdo fo the training strategy.
In the case of partitioned_training.py this is a list of batch_sizes.
When calling the training for the first time, the target set consists of a single point. Due to the training decorator, the batch_sizes get adapted to a list of ones. This however, overwrites the list in the RND agent, which will remain to be a list of ones, as it does not get updated afterwards. The training decorator only updates the batch size under certain conditions and none of them is fulfilled for batch sizes of ones.

* Adding a batch size adaption to the partitioned_training.py.
This will adapt the batch size if it was chosen too large.

* Change the train_ds_selection from Union[list, slice] to Union[np.array, slice].
This enables the usage of np when creating the indices.
Also adapting default values to that.

* Fixing the default batch size for loss_aware_reservoir.py.

* Add a test for the parameter adaption to all unit tests of training strategies.

Each training strategy has an update_training_kwargs method, which is called inside the training decorator before executing the training of a model.
The adaptions of training keywords performed in this method are tested for all training strategies.

* Adapt the test_partitioned_training_deployment.py to the use of np.arrays instead of lists.

* Adapt the notebook Using-Training-Strategies.ipynb to the use of np arrays for partitioned training.

* Update test_loss_aware_reservoir.py

Correct typo

* Update test_loss_aware_reservoir.py

change how the data set size is found out

* Define tfds version to try to fix the error.

* Define tfds version to try to fix the error.

* Define tfds version to 4.8.0 to try to fix the error.

* downgrade protobuf version to 3.20.x

* Remove unnecessary import in test_loss_aware_reservoir.py and test_partitioned_training.py

* run isort

---------

Co-authored-by: knikolaou <>